Add slug field to menu items.
As noted in #18 it's helpfull to get the slug of the referenced
object in some cases (e.g. custom routing).

Now every item and possible children have an `object_slug` field.
In case of custom links it just returns the items slug since the
id refers to the item it self.

Note that the v1 implementation isn't tested.

Closes #18