Add items argument to /menus.
In some cases the cost of fetching the menus first and then only
fetching the childs for the menus that are needed by id are higher
then fetching them all at once. It's now possible to specify an
`items` argument to indicate whether to fetch the menu items aswell
or not.

This could be also helpfull in cases like [this](https://wordpress.org/support/topic/get-menu-data-by-name-not-id) to filter on the client, although a `slug` argument to filter the menus array in addition would be need.

Note that this change is currently V2 only.